cart routing

+++ routes/cartRoutes.php
@@ -4,14 +4,10 @@ require_once '../controllers/cartController.php';
 
 $cartController = new CartController();
 
-function getCart()
-{
+if ($_SERVER['REQUEST_METHOD'] === 'GET') {
     $cartItems = $GLOBALS['cartController']->getCartItems($_SESSION['user']['id']);
     echo json_encode($cartItems);
-}
-
-function addToCart()
-{
+} else if ($_SERVER['REQUEST_METHOD'] === 'PUT') {
     if (!isset($_SESSION['user'])) {
         echo "notLoggedIn";
         exit();
@@ -20,30 +16,9 @@ function addToCart()
     $userId = $_SESSION['user']['id'];
     $cartItems = $GLOBALS['cartController']->addToCart($productId, $userId);
     echo json_encode($cartItems);
-}
-
-function deleteCartItem()
-{
+} else if ($_SERVER['REQUEST_METHOD'] === 'DELETE') {
     $productId = $_GET['id'];
     $userId = $_SESSION['user']['id'];
     return $GLOBALS['cartController']->deleteCartItem($productId, $userId);
-}
-
-
-$request = $_GET['request'];
-
-switch ($request) {
-    case 'getCartItems':
-        getCart();
-        break;
-    case 'addProduct':
-        addToCart();
-        break;
-    case 'deleteCartItem':
-        deleteCartItem();
-        break;
-    default:
-        # code...
-        break;
 }
 ?>
